implemented bruteforce calculation
updated attack API to now use BigInteger as primitive data types can't hold numbers large enough to represent our calculated values

+++ src/commonPasswords/CommonPasswordsAttack.java
@@ -5,6 +5,7 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
+import java.math.BigInteger;
 
 import types.AttackAPI;
 
@@ -69,8 +70,8 @@ public class CommonPasswordsAttack extends AttackAPI {
 	}
 	
 	@Override
-	protected int calculateMetric() {
-		return -1;
+	protected BigInteger calculateMetric() {
+		return BigInteger.valueOf(-1);
 	}
 
 	private int calculateLevenschtein(String s1, String s2) {
